[ADD] ServeContent for io.Reader
If the content's Seek method work: ServeContent uses
a seek to the end of the content to determine its size, and the param size is ignored. The same as http.ServeFile
If the content's Seek method doesn't work: ServeContent uses the param size
to generate a onlySizeSeekable as a pseudo io.ReadSeeker. If size < 0, use chunk or connection close instead